Fix C-T and C-D undo issue
Using <c-O> in the insert mapping will create a new undo block.
This mean than using <C-t> many times to reach the proper column
will need as many undo to be undone.
Using <expr> allow to do the job without exiting insert mode
(no <C-O>). However this mean that we can not run a command after
      the <C-F> to reset the restore the shift width.
This is done instead with auto command.

This also solves the issue of the cursor being move one character
     backward.